Move bot with "normal" strategy to subdir (#1)
